refactor: :art: moves BLI components to new directory
